sctp_initmsg fixes.

+++ trunk/APPS/apitester/src/sctp_utilities.h
@@ -94,20 +94,20 @@ uint32_t sctp_get_association_identifiers(int, sctp_assoc_t [], unsigned int);
 
 int 
 sctp_get_initmsg(int fd, 
-		 uint32_t *ostreams,
-		 uint32_t *istreams,
+		 uint16_t *ostreams,
+		 uint16_t *istreams,
 		 uint16_t *maxattempt,
 		 uint16_t *max_init_timeo);
 
 int 
 sctp_set_initmsg(int fd, 
-		 uint32_t ostreams,
-		 uint32_t istreams,
+		 uint16_t ostreams,
+		 uint16_t istreams,
 		 uint16_t maxattempt,
 		 uint16_t max_init_timeo);
 
-int sctp_set_im_ostream(int fd, uint32_t ostream);
-int sctp_set_im_istream(int fd, uint32_t istream);
+int sctp_set_im_ostream(int fd, uint16_t ostream);
+int sctp_set_im_istream(int fd, uint16_t istream);
 int sctp_set_im_maxattempt(int fd, uint16_t max);
 int sctp_set_im_maxtimeo(int fd, uint16_t timeo);
 int sctp_get_ndelay(int fd, uint32_t *val);
